Add files via upload
1.Interface_STK
The provided Java interface Interface_STK defines a set of methods for a stack data structure. The interface declares methods for pushing elements onto the stack (push), popping elements off the stack (pop), displaying the contents of the stack (displayStack), and checking whether the stack is full (overflow) or empty (underflow). By defining these methods in an interface, it establishes a contract that any class implementing this interface must adhere to, ensuring consistency and interoperability among different stack implementations in Java. This interface serves as a blueprint for creating stack classes with customisable behaviour while enforcing a standard set of stack operations.
2.Fixed_STK
The provided Java class Fixed_stk implements the Interface_STK interface, defining a fixed-size stack data structure. It maintains an internal array stack to store elements and a variable top to keep track of the top element's index. The constructor initialises the stack array and top index based on the provided parameters. The push method adds an element to the stack if there is space available, otherwise, it indicates that the stack is full. The pop method removes and returns the top element of the stack if it's not empty, otherwise, it indicates that the stack is empty. The displayStack method prints the elements of the stack if it's not empty. The overflow method checks if the stack is full, and the underflow method checks if the stack is empty, returning true or false accordingly. This class provides basic stack operations with fixed size, ensuring that the stack adheres to the interface's contract.
3. Growable_STk
The provided Java class Growable_stk also implements the Interface_STK interface, offering a stack implementation with dynamic growth using an ArrayList. In the constructor, it initialises the ArrayList to store elements, and top is set to -1 to indicate an empty stack. The push method adds an element to the top of the stack by incrementing top and inserting the element at that index in the ArrayList. The pop method removes and returns the top element of the stack if it's not empty, updating top accordingly. The displayStack method prints the elements of the stack. The overflow method is not applicable for a growable stack, so it prints a message and returns false. The underflow method checks if the stack is empty and returns true if top is -1, indicating an underflow condition. This class provides similar stack operations as Fixed_stk but with the flexibility of dynamically growing the stack size as needed.
4. Main
In the Main class's main method, an instance of Growable_stk named grow_stk is created. Elements 1 through 6 are then pushed onto the stack using grow_stk.push(). Finally, grow_stk.displayStack() is called to print the stack's elements.